add GoogleCloudStorageConfig::new and config and move functions to client
